batteryState.dart renamed to batter_state.dart
