+++ packages/battery/battery_platform_interface/lib/enums/battery_state.dart
